Фикс бага + перевод взаимодействий со стенами

+++ code/game/objects/structures/girders.dm
@@ -30,7 +30,7 @@
 	if(user.is_busy()) return
 	if(istype (W,/obj/item/weapon/changeling_hammer))
 		var/obj/item/weapon/changeling_hammer/C = W
-		visible_message("<span class='warning'><B>[user]</B> has punched \the <B>[src]!</B></span>")
+		visible_message("<span class='warning'><B>[src]</B> атакован <B>[user]!</B></span>")
 		user.do_attack_animation(src)
 		user.SetNextMove(CLICK_CD_MELEE)
 		if(C.use_charge(user, 1) && prob(40))
@@ -38,52 +38,52 @@
 			qdel(src)
 	else if(iswrench(W) && state == 0)
 		if(anchored && !istype(src,/obj/structure/girder/displaced))
-			to_chat(user, "<span class='notice'>Now disassembling the girder</span>")
+			to_chat(user, "<span class='notice'>Разбираем каркас...</span>")
 			if(W.use_tool(src, user, 40, volume = 100))
 				if(!src) return
-				to_chat(user, "<span class='notice'>You dissasembled the girder!</span>")
+				to_chat(user, "<span class='notice'>Вы разобрали каркас!</span>")
 				new /obj/item/stack/sheet/metal(get_turf(src))
 				qdel(src)
 		else if(!anchored)
-			to_chat(user, "<span class='notice'>Now securing the girder</span>")
+			to_chat(user, "<span class='notice'>Фиксируем каркас...</span>")
 			if(W.use_tool(src, user, 40, volume = 100))
-				to_chat(user, "<span class='notice'>You secured the girder!</span>")
+				to_chat(user, "<span class='notice'>Вы зафиксировали каркас!</span>")
 				new/obj/structure/girder( src.loc )
 				qdel(src)
 
 	else if(istype(W, /obj/item/weapon/pickaxe/plasmacutter))
-		to_chat(user, "<span class='notice'>Now slicing apart the girder</span>")
+		to_chat(user, "<span class='notice'>Режем каркас на куски</span>")
 		if(W.use_tool(src, user, 30, volume = 100))
 			if(!src) return
-			to_chat(user, "<span class='notice'>You slice apart the girder!</span>")
+			to_chat(user, "<span class='notice'>Вы разрезали каркас!</span>")
 			new /obj/item/stack/sheet/metal(get_turf(src))
 			qdel(src)
 
 	else if(istype(W, /obj/item/weapon/pickaxe/drill/diamond_drill))
-		to_chat(user, "<span class='notice'>You drill through the girder!</span>")
+		to_chat(user, "<span class='notice'>Вы сверлите сквозь каркас!</span>")
 		new /obj/item/stack/sheet/metal(get_turf(src))
 		qdel(src)
 
 	else if(isscrewdriver(W) && state == 2 && istype(src,/obj/structure/girder/reinforced))
-		to_chat(user, "<span class='notice'>Now unsecuring support struts</span>")
+		to_chat(user, "<span class='notice'>Ослабляем поддерживающие опоры...</span>")
 		if(W.use_tool(src, user, 40, volume = 100))
 			if(!src) return
-			to_chat(user, "<span class='notice'>You unsecured the support struts!</span>")
+			to_chat(user, "<span class='notice'>Вы ослабили поддерживающие опоры!</span>")
 			state = 1
 
 	else if(iswirecutter(W) && istype(src,/obj/structure/girder/reinforced) && state == 1)
-		to_chat(user, "<span class='notice'>Now removing support struts</span>")
+		to_chat(user, "<span class='notice'>Удаляем поддерживающие опоры...</span>")
 		if(W.use_tool(src, user, 40, volume = 100))
 			if(!src) return
-			to_chat(user, "<span class='notice'>You removed the support struts!</span>")
+			to_chat(user, "<span class='notice'>Вы удалили поддерживающие опоры!</span>")
 			new/obj/structure/girder( src.loc )
 			qdel(src)
 
 	else if(iscrowbar(W) && state == 0 && anchored )
-		to_chat(user, "<span class='notice'>Now dislodging the girder</span>")
+		to_chat(user, "<span class='notice'>Смещаем каркас...</span>")
 		if(W.use_tool(src, user, 40, volume = 100))
 			if(!src) return
-			to_chat(user, "<span class='notice'>You dislodged the girder!</span>")
+			to_chat(user, "<span class='notice'>Вы сместили каркас!</span>")
 			new/obj/structure/girder/displaced( src.loc )
 			qdel(src)
 
@@ -96,15 +96,15 @@
 				if(!anchored)
 					if(!S.use(2))
 						return
-					to_chat(user, "<span class='notice'>You create a false wall! Push on it to open or close the passage.</span>")
+					to_chat(user, "<span class='notice'>Вы сделали потайную дверь! Толкните чтобы открыть или закрыть проход.</span>")
 					new /obj/structure/falsewall (src.loc)
 					qdel(src)
 				else
 					if(S.get_amount() < 2)
 						return ..()
-					to_chat(user, "<span class='notice'>Now adding plating...</span>")
+					to_chat(user, "<span class='notice'>Добавляем металл...</span>")
 					if(S.use_tool(src, user, 40, amount = 2, volume = 100))
-						to_chat(user, "<span class='notice'>You added the plating!</span>")
+						to_chat(user, "<span class='notice'>Вы добавили металл!</span>")
 						var/turf/Tsrc = get_turf(src)
 						Tsrc.ChangeTurf(/turf/simulated/wall)
 						for(var/turf/simulated/wall/X in Tsrc.loc)
@@ -116,16 +116,16 @@
 				if(!anchored)
 					if(!S.use(2))
 						return
-					to_chat(user, "<span class='notice'>You create a false wall! Push on it to open or close the passage.</span>")
+					to_chat(user, "<span class='notice'>Вы сделали потайную дверь! Толкните чтобы открыть или закрыть проход.</span>")
 					new /obj/structure/falsewall/reinforced(loc)
 					qdel(src)
 				else
 					if (istype (src, /obj/structure/girder/reinforced))
 						if(S.get_amount() < 1)
 							return ..()
-						to_chat(user, "<span class='notice'>Now finalising reinforced wall.</span>")
+						to_chat(user, "<span class='notice'>Завершаем укрепленную стену....</span>")
 						if(S.use_tool(src, user, 50, amount = 1, volume = 100))
-							to_chat(user, "<span class='notice'>Wall fully reinforced!</span>")
+							to_chat(user, "<span class='notice'>Стена полностью укреплена!</span>")
 							var/turf/Tsrc = get_turf(src)
 							Tsrc.ChangeTurf(/turf/simulated/wall/r_wall)
 							for(var/turf/simulated/wall/r_wall/X in Tsrc.loc)
@@ -135,28 +135,30 @@
 					else
 						if(S.get_amount() < 1)
 							return ..()
-						to_chat(user, "<span class='notice'>Now reinforcing girders</span>")
+						to_chat(user, "<span class='notice'>Укрепляем каркас...</span>")
 						if(S.use_tool(src, user, 60, amount = 1, volume = 100))
-							to_chat(user, "<span class='notice'>Girders reinforced!</span>")
+							to_chat(user, "<span class='notice'>Каркас укреплен!</span>")
 							new/obj/structure/girder/reinforced( src.loc )
 							qdel(src)
 						return
 
 		if(S.sheettype)
 			var/M = S.sheettype
+			if (!(M in list("gold", "diamond", "uranium", "silver", "sandstone", "phoron")))
+				return
 			if(!anchored)
 				if(!S.use(2))
 					return
-				to_chat(user, "<span class='notice'>You create a false wall! Push on it to open or close the passage.</span>")
+				to_chat(user, "<span class='notice'>Вы сделали потайную дверь! Толкните чтобы открыть или закрыть проход.</span>")
 				var/F = text2path("/obj/structure/falsewall/[M]")
 				new F (src.loc)
 				qdel(src)
 			else
 				if(S.get_amount() < 2)
 					return ..()
-				to_chat(user, "<span class='notice'>Now adding plating...</span>")
+				to_chat(user, "<span class='notice'>Добавляем материал...</span>")
 				if(S.use_tool(src, user, 40, amount = 2, volume = 100))
-					to_chat(user, "<span class='notice'>You added the plating!</span>")
+					to_chat(user, "<span class='notice'>Вы добавили материал!</span>")
 					var/turf/Tsrc = get_turf(src)
 					Tsrc.ChangeTurf(text2path("/turf/simulated/wall/mineral/[M]"))
 					for(var/turf/simulated/wall/mineral/X in Tsrc.loc)
@@ -171,7 +173,7 @@
 		if (P.pipe_type in list(0, 1, 5))	//simple pipes, simple bends, and simple manifolds.
 			user.drop_item()
 			P.loc = src.loc
-			to_chat(user, "<span class='notice'>You fit the pipe into the [src]!</span>")
+			to_chat(user, "<span class='notice'>Вы встроили трубу в каркас!</span>")
 	else
 		..()
 
@@ -204,9 +206,9 @@
 /obj/structure/girder/attack_animal(mob/living/simple_animal/attacker)
 	if(attacker.environment_smash)
 		..()
-		attacker.visible_message("<span class='warning'>[attacker] smashes against [src].</span>", \
-			 "<span class='warning'>You smash against [src].</span>", \
-			 "You hear twisting metal.")
+		attacker.visible_message("<span class='warning'>[attacker] крушит каркас.</span>", \
+			 "<span class='warning'>Вы крушите каркас.</span>", \
+			 "Вы слышите скрежет металла.")
 		playsound(src, 'sound/effects/grillehit.ogg', VOL_EFFECTS_MASTER)
 		health -= attacker.melee_damage
 		if(health <= 0)
@@ -239,21 +241,21 @@
 	if(user.is_busy(src))
 		return
 	if(iswrench(W))
-		to_chat(user, "<span class='notice'>Now disassembling the girder</span>")
+		to_chat(user, "<span class='notice'>Разбираем каркас...</span>")
 		if(W.use_tool(src, user, 40, volume = 100))
-			to_chat(user, "<span class='notice'>You dissasembled the girder!</span>")
+			to_chat(user, "<span class='notice'>Вы разобрали каркас!</span>")
 			new /obj/effect/decal/remains/human(get_turf(src))
 			qdel(src)
 
 	else if(istype(W, /obj/item/weapon/pickaxe/plasmacutter))
-		to_chat(user, "<span class='notice'>Now slicing apart the girder</span>")
+		to_chat(user, "<span class='notice'>Режем каркас на куски...</span>")
 		if(W.use_tool(src, user, 30, volume = 100))
-			to_chat(user, "<span class='notice'>You slice apart the girder!</span>")
+			to_chat(user, "<span class='notice'>Вы разрезали каркас!</span>")
 		new /obj/effect/decal/remains/human(get_turf(src))
 		qdel(src)
 
 	else if(istype(W, /obj/item/weapon/pickaxe/drill/diamond_drill))
-		to_chat(user, "<span class='notice'>You drill through the girder!</span>")
+		to_chat(user, "<span class='notice'>Вы сверлите сквозь каркас!</span>")
 		new /obj/effect/decal/remains/human(get_turf(src))
 		qdel(src)
